requirements.txt
cesium>=0.6.1
netCDF4
peewee>=2.8.3
psycopg2
requests>=2.6.0
pyyaml
tornado
pyzmq
pyjwt
plotly
simplejson
distributed
selenium
pytest
